add cpu weight for n2_quant
