+++ scripts/n2.py
@@ -6,6 +6,8 @@ from typing import Any, List, Tuple
 from transformers import (PreTrainedTokenizerBase, AutoTokenizer)
 import random
 import datasets
+from vllm.utils import reset_seed
+reset_seed()
 # get file location
 file_path = os.path.abspath(__file__)
 dataset_path = os.path.join(os.path.dirname(file_path), "../benchmarks")
