+++ scripts/n2_quant.py
@@ -6,6 +6,8 @@ from typing import Any, List, Tuple
 from transformers import (PreTrainedTokenizerBase, AutoTokenizer)
 import random
 import datasets
+from vllm.utils import reset_seed
+reset_seed()
 # get file location
 file_path = os.path.abspath(__file__)
 dataset_path = os.path.join(os.path.dirname(file_path), "../benchmarks")
@@ -185,6 +187,7 @@ if __name__ == "__main__":
         distributed_executor_backend='ray',
         trust_remote_code=True,
         quantization='inc_q',
+        weights_load_device="cpu",
         max_model_len=16384,
         dtype="bfloat16",
     )
